converted GT-11270 to enyo 2.6
