+++ test/moonstone/TimePicker/GT-11270-DaylightSavings/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "GT-11270-DaylightSavings.js"
+}
